project structure fixed
