+++ src/components/TodoList/TodoList.js
@@ -1,22 +1,12 @@
 import { useState, useEffect } from "react";
 import "./TodoList.css";
 import InputForm from "../InputForm/InputForm";
-import TopBar from "../TopBar/TopBar";
 
-const TodoList = () => {
+const TodoList = ({ search }) => {
   const [todos, setTodos] = useState([]);
   const [isComponentLoaded, setIsComponentLoaded] = useState(false);
   const [remainingItems, setRemainingItems] = useState(0);
-  const [text, setText] = useState({
-    id: 1,
-    input: "",
-    isChecked: false,
-  });
-
-  const [search, setSearch] = useState("");
-  const handleSearch = (term) => {
-    setSearch(term);
-  };
+  const searchText = search || "";
 
   useEffect(() => {
     const storedTodos = JSON.parse(localStorage.getItem("todos") || "[]");
@@ -31,7 +21,7 @@ const TodoList = () => {
     const remainingTodos = todos.filter((todo) => !todo.isChecked).length;
 
     setRemainingItems(remainingTodos);
-  }, [text.isChecked, todos, isComponentLoaded]);
+  }, [todos, isComponentLoaded]);
 
   if (!isComponentLoaded) {
     return <h2>Loading...</h2>; //add spinner
@@ -40,15 +30,15 @@ const TodoList = () => {
   //todos array filter
   const filteredTodos = todos.filter(
     (todo) =>
-      todo.input && todo.input.toLowerCase().includes(search.toLowerCase())
+      (todo.input &&
+        todo.input.toLowerCase().includes(searchText.toLowerCase())) ||
+      false
   );
 
   return (
     <div>
       <h3>REMAINING TASKS: {remainingItems}</h3>
 
-      {/* rendering TopBar with props */}
-      <TopBar placeholder="Search for a task" onSearch={handleSearch} />
       {/* passing filtered Todos list */}
       <InputForm todos={filteredTodos} setTodos={setTodos} />
