+++ src/pages/MainPage.js
@@ -1,12 +1,20 @@
+import React, { useState } from "react";
 import MainContent from "../components/MainContent/MainContent";
 import TopBar from "../components/TopBar/TopBar";
 import "./MainPage.css";
 
 const MainPage = () => {
+  const [search, setSearch] = useState("");
+
+  const handleSearch = (term) => {
+    setSearch(term);
+  };
+
   return (
     <>
-      <TopBar />
-      <MainContent />
+      {/* rendering TopBar with props */}
+      <TopBar placeholder="Search for a task" onSearch={handleSearch} />
+      <MainContent search={search} />
     </>
   );
 };
